Spatial Graphs with traditional controls
This is an initial commit.

Here we have a canvas that lets user plot
random points and join two points with a segment. The points are
generated randomly and if a point already exists at the given position,
the button simply does nothing. The same is true for segments.

Additionally, this commit also gives the user to remove the points and
segments or clear the canvas completely. Deleting a point also removes
all the associated segments, as having segments without points would
make no sense.