Improve graph editor with modern implementation
This commit introduces modern interaction system to the graph editor.
Previously, there was no graph editor, just canvas and buttons were used
to add and remove points and segments randomly. This approach was not
suitable for the final product.

Now, a graph editor is introduced that lets the user add points and
segments using mouse clicks. The user can select points, and drag it
wherever they desire. They can unselect those points and segments are
created when a point is already selected.

Overall, the user interface is now more interactive and easy to use.